agent-stack-oss/requirements.txt

langchain>=0.2
langgraph>=0.2
pydantic>=2.7
python-dotenv>=1.0
llama-index>=0.10
chromadb>=0.5
duckdb>=1.0
openai>=1.40
tiktoken>=0.7
httpx>=0.27
ollama>=0.3
langfuse>=2.33
arize-phoenix>=7.0
g4f>=0.3.3
pytest>=8.3
mypy>=1.11
ruff>=0.6
